# zhiliao-ainame/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/code": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Email Code",
        "operationId": "get_email_code_auth_code_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/name/": {
      "post": {
        "summary": "Take Names",
        "operationId": "take_names_name__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NameOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/hello/{name}": {
      "get": {
        "summary": "Say Hello",
        "operationId": "say_hello_hello__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/test": {
      "get": {
        "summary": "Mail Test",
        "operationId": "mail_test_mail_test_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Password",
            "description": "密码"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn"
      },
      "LoginOut": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "user",
          "token"
        ],
        "title": "LoginOut"
      },
      "NameIn": {
        "properties": {
          "surname": {
            "type": "string",
            "title": "Surname",
            "description": "姓氏"
          },
          "gender": {
            "type": "string",
            "enum": [
              "不限",
              "男",
              "女"
            ],
            "title": "Gender",
            "description": "性别"
          },
          "length": {
            "type": "string",
            "enum": [
              "不限",
              "单字",
              "两字"
            ],
            "title": "Length",
            "description": "字数"
          },
          "other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other",
            "description": "其他要求",
            "default": ""
          },
          "exclude": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude",
            "description": "排除的名字",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "surname",
          "gender",
          "length"
        ],
        "title": "NameIn"
      },
      "NameOut": {
        "properties": {
          "names": {
            "items": {
              "$ref": "#/components/schemas/NameSchema"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "type": "object",
        "required": [
          "names"
        ],
        "title": "NameOut"
      },
      "NameSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "姓名"
          },
          "reference": {
            "type": "string",
            "title": "Reference",
            "description": "出处"
          },
          "moral": {
            "type": "string",
            "title": "Moral",
            "description": "寓意"
          }
        },
        "type": "object",
        "required": [
          "name",
          "reference",
          "moral"
        ],
        "title": "NameSchema"
      },
      "RegisterIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Username",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Password",
            "description": "密码"
          },
          "confirm_password": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Confirm Password",
            "description": "密码"
          },
          "code": {
            "type": "string",
            "maxLength": 4,
            "minLength": 4,
            "title": "Code",
            "description": "邮箱验证码"
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "password",
          "confirm_password",
          "code"
        ],
        "title": "RegisterIn"
      },
      "ResponseOut": {
        "properties": {
          "result": {
            "type": "string",
            "enum": [
              "success",
              "failure"
            ],
            "title": "Result",
            "description": "操作的结果！",
            "default": "success"
          }
        },
        "type": "object",
        "title": "ResponseOut",
        "description": "用于一些视图函数，只要返回操作结果的模型"
      },
      "UserSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Username",
            "description": "用户名"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "username"
        ],
        "title": "UserSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}